pyproject.toml
[tool.poetry]
name = "foodies"
version = "0.1.0"
description = ""
authors = ["fatemahAlMomen <fatemah.alm@hotmail.com>"]

[tool.poetry.scripts]
manage = "manage:main"

[tool.poetry.dependencies]
python = "^3.9.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
